Rename test scripts
